# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mural Maz Lince API",
    "description": "API REST del sistema Mural Maz Lince. Digitaliza el tablero de anuncios estudiantil de la Universidad Autónoma de Occidente, Unidad Regional Mazatlán.\n\n**Estado:** Definitivo Sprint 10 — incluye pipeline de moderación (Vision + bad-words + dHash), panel admin y endpoints operativos de shadowban/cola.\n\n**Autenticación:** Cookies HttpOnly con JWT (ADR-04). Incluir `credentials: 'include'` en cada solicitud del cliente.\n\n**Referencia SRS:** v1.4 | **Referencia SAD:** v1.3 | **Referencia DDC:** v1.2",
    "version": "1.0.0-sprint10",
    "contact": {
      "name": "Miriam Alanis — Directora del Proyecto",
      "url": "https://github.com/devHamialco/MuralMazLince"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Entorno local (docker-compose)"
    },
    {
      "url": "https://muralmaz.railway.app",
      "description": "Producción (Railway.app) — disponible al cierre del Sprint 14"
    }
  ],
...
  "paths": {
...
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "Listar proyectos del emprendedor",
        "description": "Devuelve todos los proyectos (activos y suspendidos) del emprendedor autenticado con sus anuncios activos y métricas resumidas. **RF-15, RF-23.**",
        "operationId": "getProjects",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de proyectos del emprendedor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Project" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Crear nuevo proyecto",
        "description": "Crea un proyecto para el emprendedor autenticado. Valida que no se supere el límite de proyectos activos simultáneos (MAX_ACTIVE_PROJECTS_PER_USER, default: 5). **RF-15, RN-03.**",
        "operationId": "createProject",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectRequest" },
              "example": {
                "name": "Tech Lince — Desarrollo Web",
                "description": "Servicios de desarrollo web y apps para la comunidad universitaria.",
                "category_id": 11
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proyecto creado correctamente.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "tags": ["projects"],
        "summary": "Detalle de un proyecto",
        "description": "Devuelve el detalle completo de un proyecto con su lista de anuncios (activos e historial) y métricas de impacto filtradas por umbral de intención. **RF-16, RF-23, RF-24, RN-04.**",
        "operationId": "getProjectById",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "responses": {
          "200": {
            "description": "Detalle del proyecto.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "patch": {
        "tags": ["projects"],
        "summary": "Editar proyecto",
        "description": "Edita el nombre, descripción o categoría de un proyecto del emprendedor autenticado. **RF-16.**",
        "operationId": "updateProject",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proyecto actualizado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Eliminar proyecto",
        "description": "Elimina definitivamente un proyecto y todos sus anuncios activos con sus imágenes en Cloudinary. Operación irreversible. **RF-18, RN-06.**",
        "operationId": "deleteProject",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "responses": {
          "200": {
            "description": "Proyecto y anuncios asociados eliminados correctamente.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "message": "Proyecto eliminado correctamente." }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/projects/{id}/status": {
      "patch": {
        "tags": ["projects"],
        "summary": "Cambiar estado del proyecto (activo ↔ suspendido)",
        "description": "Alterna el estado del proyecto entre `active` y `suspended`. Un proyecto suspendido no muestra sus anuncios en el feed pero conserva su historial. **RF-17, RN-06.**",
        "operationId": "updateProjectStatus",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectStatusRequest" },
              "example": { "status": "suspended" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado del proyecto actualizado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT en cookie HttpOnly. Emitido por POST /auth/login. No accesible desde JavaScript del cliente (ADR-04). Incluir `credentials: 'include'` en cada solicitud fetch del frontend."
      }
    },
...
    "parameters": {
...
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID del proyecto.",
        "schema": { "type": "integer", "example": 7 }
      },
...
    },
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Descripción del error. Nunca expone datos sensibles." }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "code": { "type": "string", "example": "CAT-01" },
          "name": { "type": "string", "example": "Alimentos y bebidas" },
          "group_name": { "type": "string", "example": "Productos físicos" },
          "is_active": { "type": "boolean", "example": true }
        }
      },
      "AnnouncementSummary": {
        "type": "object",
        "description": "Vista resumida del anuncio para el feed. RF-08.",
        "properties": {
          "id": { "type": "integer", "example": 42 },
          "title": { "type": "string", "example": "Diseño de logos profesionales" },
          "cloudinary_url": { "type": "string", "format": "uri", "example": "https://res.cloudinary.com/mural/image/upload/v1/mural_lince/abc123.jpg" },
          "category": { "$ref": "#/components/schemas/Category" },
          "project_name": { "type": "string", "example": "Tech Lince — Diseño Gráfico" },
          "entrepreneur_display_name": {
            "type": "string",
            "description": "Proveniente de entrepreneur_profiles.display_name. RF-08, RF-39.",
            "example": "Mariana G. Tech Lince"
          },
          "total_likes": { "type": "integer", "example": 14 },
          "average_rating": { "type": "number", "format": "float", "example": 2.7 },
          "total_ratings": { "type": "integer", "example": 11 },
          "expires_at": { "type": "string", "format": "date-time", "example": "2026-05-15T23:59:59Z" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-04-10T14:30:00Z" }
        }
      },
      "AnnouncementDetail": {
        "type": "object",
        "description": "Vista completa del anuncio. Incluye contact_link solo para usuarios autenticados. RF-10, RF-11.",
        "allOf": [
          { "$ref": "#/components/schemas/AnnouncementSummary" },
          {
            "type": "object",
            "properties": {
              "description": {
                "type": "string",
                "maxLength": 500,
                "example": "Diseño de logotipos, tarjetas de presentación y material visual para tu emprendimiento."
              },
              "status": {
                "type": "string",
                "enum": ["active", "pending_review", "rejected", "shadowban"],
                "example": "active"
              },
              "contact_link": {
                "type": "string",
                "format": "uri",
                "nullable": true,
                "description": "URL wa.me generada dinámicamente en el servidor. null para ROL-01. El número nunca se expone. RF-11, RN-09.",
                "example": "https://wa.me/5216691234567"
              },
              "project_id": { "type": "integer", "example": 7 },
              "custom_category": {
                "type": "string",
                "nullable": true,
                "description": "Solo presente si category.code = CAT-00 (Otro). RN-12.",
                "example": null
              }
            }
          }
        ]
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 7 },
          "name": { "type": "string", "example": "Tech Lince — Desarrollo Web" },
          "description": { "type": "string", "example": "Servicios de desarrollo web para la comunidad universitaria." },
          "category": { "$ref": "#/components/schemas/Category" },
          "status": { "type": "string", "enum": ["active", "suspended"], "example": "active" },
          "active_announcements_count": { "type": "integer", "example": 2 },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProjectDetail": {
        "type": "object",
        "description": "Detalle completo de proyecto con anuncios y métricas. RF-16, RF-23, RF-24.",
        "allOf": [
          { "$ref": "#/components/schemas/Project" },
          {
            "type": "object",
            "properties": {
              "announcements": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AnnouncementDetail" }
              },
              "metrics": { "$ref": "#/components/schemas/ProjectMetrics" },
              "active_announcements_limit": {
                "type": "integer",
                "description": "Límite máximo de anuncios activos por proyecto (MAX_ACTIVE_ANNOUNCEMENTS_PER_PROJECT). RN-04.",
                "example": 3
              }
            }
          }
        ]
      },
      "ProjectMetrics": {
        "type": "object",
        "description": "Métricas de impacto filtradas por umbral de intención. RF-24, RF-14, RN-08.",
        "properties": {
          "total_valid_likes": {
            "type": "integer",
            "description": "Likes con is_accidental = false.",
            "example": 87
          },
          "average_rating": {
            "type": "number",
            "format": "float",
            "description": "Promedio de valoraciones con is_accidental = false.",
            "example": 2.6
          },
          "total_views": { "type": "integer", "example": 430 }
        }
      },
      "CreateProjectRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 120, "example": "Tech Lince — Desarrollo Web" },
          "description": { "type": "string" },
          "category_id": { "type": "integer", "nullable": true, "example": 11 }
        }
      },
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 120 },
          "description": { "type": "string" },
          "category_id": { "type": "integer", "nullable": true }
        }
      },
      "UpdateProjectStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["active", "suspended"],
            "description": "Nuevo estado del proyecto. RN-06."
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "400 Bad Request — Checksum inválido, campos faltantes o formato de WhatsApp incorrecto.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "El formato de la matrícula es inválido. Debe tener 8 dígitos y comenzar con 2." }
          }
        }
      },
      "Unauthorized": {
        "description": "401 Unauthorized — Cookie de sesión ausente o expirada.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Sesión no encontrada. Por favor inicia sesión." }
          }
        }
      },
      "Forbidden": {
        "description": "403 Forbidden — Cuenta suspendida, o rol insuficiente para la acción solicitada.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Tu cuenta está suspendida. Contacta al administrador." }
          }
        }
      },
      "NotFound": {
        "description": "404 Not Found — El recurso solicitado no existe.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Anuncio no encontrado." }
          }
        }
      },
      "Conflict": {
        "description": "409 Conflict — Matrícula ya registrada, reporte duplicado o límite de proyectos/anuncios alcanzado.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Esta matrícula ya está registrada en el sistema." }
          }
        }
      },
...
      "InternalServerError": {
        "description": "500 Internal Server Error — Error inesperado en el servidor. El detalle se registra en el log; el cliente recibe un mensaje genérico.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Ocurrió un error interno. Por favor intenta más tarde." }
          }
        }
      }
    }
  }
}